Add a simple test that really just calls `percy snapshot` on an html file.
this html file was originally snapshotted while it was linked to the standard bootstrap.css file, making it a master file of sorts. Then the css linking was changed to the uselesscss version of the bootstrap file.

Update .travis to ensure ruby environment is late enough

Replace unit test with just a CLI call from the CI